# docs/swagger.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Mini E-Commerce API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Ordenes": {
      "post": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Crea una nueva orden con los items del carrito del usuario.",
        "description": "Requiere autenticación.\r\n\r\n\r\n\r\nReglas de negocio:\r\n- El carrito debe existir\r\n- El carrito no puede estar vacío\r\n- Todos los productos deben existir y estar activos\r\n- Debe haber stock suficiente\r\n\r\n\r\n\r\nLa orden se crea con estado <b>Pending</b>.",
        "responses": {
          "200": {
            "description": "Orden creada correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "usuarioId inválido en el token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Carrito no existe o producto no existe.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Carrito vacío o stock insuficiente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Obtiene las órdenes del usuario autenticado.",
        "description": "Requiere autenticación.\r\nDevuelve todas las órdenes asociadas al usuario.",
        "responses": {
          "200": {
            "description": "Listado de órdenes.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoListApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "usuarioId inválido en el token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ordenes/{ordenId}": {
      "get": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Obtiene el detalle de una orden específica del usuario.",
        "description": "Requiere autenticación.\r\nSolo devuelve la orden si pertenece al usuario autenticado.",
        "parameters": [
          {
            "name": "ordenId",
            "in": "path",
            "description": "Id de la orden.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la orden.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdenDtoApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ordenId inválido o usuarioId inválido en token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Orden no existe o no pertenece al usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ordenes/cancelar/{ordenId}": {
      "post": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Cancela una orden pendiente del usuario.",
        "description": "Requiere autenticación.\r\n\r\n\r\n\r\nReglas:\r\n- Solo se puede cancelar si la orden está en estado <b>Pending</b>\r\n- No se puede cancelar si ya está cancelada\r\n- No se puede cancelar si está pagada u otro estado",
        "parameters": [
          {
            "name": "ordenId",
            "in": "path",
            "description": "Id de la orden.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden cancelada correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ordenId inválido o usuarioId inválido en token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Orden no existe o no pertenece al usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: ya cancelada o estado inválido para cancelar.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ordenes/pagar/{ordenId}": {
      "post": {
        "tags": [
          "Ordenes"
        ],
        "summary": "Paga una orden pendiente del usuario.",
        "description": "Requiere autenticación.\r\n\r\n\r\n\r\nReglas:\r\n- Solo se puede pagar si la orden está en <b>Pending</b>\r\n- Valida stock antes de pagar\r\n- Descuenta stock\r\n- Vacía el carrito\r\n- Cambia estado a <b>Paid</b>",
        "parameters": [
          {
            "name": "ordenId",
            "in": "path",
            "description": "Id de la orden.",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orden pagada correctamente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "ordenId inválido o usuarioId inválido en token.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Orden no existe o no pertenece al usuario.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Conflicto: estado inválido o stock insuficiente.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectApiResponseDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ObjectApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "OrdenDetallesDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "orderId": {
            "type": "integer",
            "format": "int32"
          },
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "OrdenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "format": "int32"
          },
          "totalAmount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/StatusOrden"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "ordenDetallesDtos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenDetallesDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "OrdenDtoApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "$ref": "#/components/schemas/OrdenDto"
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
      "OrdenDtoListApiResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrdenDto"
            },
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          }
        },
        "additionalProperties": false
      },
...
      "StatusOrden": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese: Bearer {tu_token}",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}